perf(rocm): reduce Kimi-K3 fallback decode staging
Write packed recurrent KDA decode directly into a validated caller-owned output for pure non-speculative fallback batches while retaining allocated placement for mixed paths.

Keep the MoE correction bias in FP32 to match router logits and avoid a per-forward grouped-top-k cast. Keep AMD and NVIDIA recurrent-wrapper contracts aligned and cover output aliasing, state equality, padding, and invalid buffers.

Generated-by: OpenAI Codex

+++ vllm/models/kimi_k3/amd/linear.py
@@ -210,7 +210,10 @@ class KimiMoE(nn.Module):
             prefix=f"{prefix}.gate",
         )
 
-        self.gate.e_score_correction_bias = nn.Parameter(torch.empty(num_experts))
+        # Match FP32 router logits to avoid a per-forward grouped-top-k cast.
+        self.gate.e_score_correction_bias = nn.Parameter(
+            torch.empty(num_experts, dtype=torch.float32)
+        )
 
         if self.num_shared_experts is not None:
             shared_intermediate_size = moe_intermediate_size * self.num_shared_experts
